--- .github/workflows/tmp-source-provenance-hardening.yml
@@ -1,26 +1,0 @@
-name: Temporary source provenance hardening
-
-on:
-  push:
-    branches: [feature/source-provenance-receipts]
-
-permissions:
-  contents: write
-
-jobs:
-  harden:
-    runs-on: ubuntu-24.04
-    steps:
-      - uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1
-      - run: python -m pip install -e '.[test]'
-      - run: python scripts/tmp_harden_source_provenance.py
-      - run: python -m pytest tests/unit/test_source_provenance.py -q
-      - run: |
-          rm scripts/tmp_harden_source_provenance.py
-          rm .github/workflows/tmp-source-provenance-hardening.yml
-          git config user.name VESSAXOR
-          git config user.email vessaxor@gmail.com
-          git add -A
-          git diff --cached --check
-          git commit -m "Harden source authorization witness verification"
-          git push origin HEAD:feature/source-provenance-receipts
